+++ CHANGES.rst
@@ -11,6 +11,7 @@ CHANGELOG
 - allow to match multiple attributes against a device (AND-wise)
 - allow to overwrite udiskies default handleability settings
 - raise exception if --config file doesn't exist
+- add --options parameter for udiskie-mount
 
 
 0.7.0